requirements.txt
bcrypt==4.0.1
blinker==1.6.2
boto3==1.28.21
botocore==1.31.20
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
Flask==2.3.2
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.3
psycopg2==2.9.7
python-dateutil==2.8.2
requests==2.31.0
s3transfer==0.6.1
six==1.16.0
urllib3==2.0.4
Werkzeug==2.3.6
zipp==3.16.2
python-dotenv~=1.0.0